Moved setEnvVars() to routes-utils.js, tested
